pyproject.toml
[tool.poetry]
name = "konnector"
version = "0.1.0"
description = ""
authors = ["Jimbroze <james.n.dickinson@gmail.com>"]
packages = [{ include = "konnector", from = "src" }]


[tool.poetry.dependencies]
python = "~3.9"
Flask = "^2.0.2"
python-dateutil = "^2.8.2"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
APScheduler = "^3.9.1"
pytest-lazy-fixture = "^0.6.3"
attrs = "^23.1.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pytest = "^7.2.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = ["src"]